snap

+++ t
@@ -10,16 +10,6 @@ fi
 input=$filename
 filename_withoutext=${filename%.*}
 
-#n=1&&ss=0&&S1="文在寅访华丧权辱国 受尽屈辱 因传播郭文贵爆料 被捕判刑者激增 《今日热评》12.20"
-#n=3&&ss=25&&S1="又是減稅又是 廢除網絡中立 川普到底靠不靠譜？ 《明鏡專訪》12.20"
-#n=2&&ss=0&&S1="陳破空文昭对谈 拆毛邓题词不拆江题词 习近平有何盘算 新唐人电视 2017.12.19"
-#n=3&&ss=8&&S1="著名人權律師劉士輝 為計生受害者打官司 《計劃生育回頭看》第6期 2017.12.20"
-#n=1&&ss=0&&S1="【环球直击】 12月20日完整版 新唐人电视 2017.12.20"
-#n=1&&ss=0&&S1="黄琦病危 下月移送法院审讯 新唐人电视 2017.12.20"
-#n=1&&ss=0&&S1="沧县天然气爆炸伤人 民众市府抗议 新唐人电视 2017.12.20"
-#n=1 && ss=0 && S1="這三個月不會打朝鮮 《點點今天事》 2017年12月20日"
-#n=4 && ss=41 && S1="蝴蝶三人行 盲流子曝张欣平语音 揭露其欺骗网友行径 2017.12.19"
-
 datefull=`TZ="UTC-8" date "+%y%m%d%H%M%S"`
 DURATION_HMS=$(ffmpeg -i "$input" 2>&1 | grep Duration | cut -f 4 -d ' ')
 DURATION_H=$(echo "$DURATION_HMS" | cut -d ':' -f 1)
@@ -36,6 +26,10 @@ then
 	v=300
 fi
 
+if [ "$volume" = "" ]; then
+	volume=1
+fi
+
 tech="ffmpeg 加速${speed} 音40k 视${v}k ${fps}fps $datefull"
 
 if [ -f sub/${filename_withoutext}.ass ]
@@ -45,7 +39,7 @@ then
 	input=$new
 fi
 
-if [ $ss != 0 ]
+if [ "$ss" != "0" ]
 then
 	new=${filename_withoutext}_ss.mkv
 	ffmpeg -y -i $input -vcodec copy -acodec copy -ss $ss $new
@@ -66,7 +60,7 @@ then
 	-i $input -y \
       	-strict -2 -q:a 1.5 -ac 1 -ar 8000 \
        	-b:v ${v}k -r 25 \
-       	-filter_complex "[0:v]setpts=PTS/${speed}[v];[0:a]atempo=${speed}[a]" \
+       	-filter_complex "[0:v]setpts=PTS/${speed}[v];[0:a]volume=$volume[af];[af]atempo=${speed}[a]" \
        	-map "[v]" -map "[a]" \
 	-f segment -segment_time $len -reset_timestamps 1 \
        	${filename_withoutext}_%d.mp4
@@ -93,10 +87,13 @@ else
 	ffmpeg \
 	-i $input -y \
       	-strict -2 -q:a 1.5 -ac 1 -ar 8000 \
-       	-b:v ${v}k -r 25 \
-       	-filter_complex "[0:v]setpts=PTS/${speed}[v1];[v1]ass=$filename_withoutext.ass[v];[0:a]highpass=f=200,lowpass=f=3000[af];[af]atempo=${speed}[a]" \
+     	-b:v ${v}k -r ${fps} \
+       	-filter_complex "[0:v]setpts=PTS/${speed}[v1];[v1]ass=$filename_withoutext.ass[v];[0:a]volume=$volume[af];[af]atempo=${speed}[a]" \
        	-map "[v]" -map "[a]" \
     	/home/public_share/${datefull}_${filename_withoutext}_${speed}_$a$v${fps}.mp4
 fi
-rm -f $filename_withoutext.ass
-rm -f ${filename_withoutext}_ss.*
+
+if [ "a$2" == "a-q" ]; then
+	rm -f $filename_withoutext.ass
+	rm -f ${filename_withoutext}_ss.*
+fi
